Move congress lookup to its own file. Keep phone controller strictly route handlers.
